♻️ refactor InBetweenButton
