+++ src/pages/BonusQuestion.tsx
@@ -5,6 +5,7 @@ import React, { useEffect, useMemo } from 'react'
 import { useHistory } from 'react-router'
 
 import MovieCard from '../components/card/MovieCard'
+import { InBetweenButton } from '../components/InBetweenButton'
 import useStore from '../useStore'
 import { MovieData, MOVIES } from './Question'
 
@@ -16,23 +17,6 @@ const GET_BONUS_SCORE = gql`
     )
   }
 `
-interface InBetweenButtonProps {
-  ix: number
-  scoreQuestion: (ix: number) => void
-}
-
-const InBetweenButton = ({ scoreQuestion, ix }: InBetweenButtonProps) => {
-  return (
-    <div className="absolute" style={{ top: '-2em' }}>
-      <button
-        className="text-xs bg-teal-500 hover:bg-teal-400 text-white font-bold py-3 px-4 border-b-4 border-teal-700 hover:border-teal-500 rounded-full shadow-sm"
-        onClick={() => scoreQuestion(ix)}
-      >
-        HERE
-      </button>
-    </div>
-  )
-}
 
 /**
  * ### The screen where the user can double his/her points.
